Moved enums & support classes from solitaire.h to solitaire_common.h. Cleaned up imports.
